Update flag triggers/sets/unsets
Changed some stuff around so that if you go down the Trainer Tower Elevator without speaking to the Emerald giver, the elevator will NOT malfunction. It will only malfunction if you have the Emerald flag set. The Emerald giver will only give you the Emerald if you don't have one yet. So, when the Emerald flag is cleared in the elevator, it will not malfunction again, yet the giver will not give another Emerald.